fix: resolve build errors for static export
