+++ src/app/admin/courses/[id]/layout.tsx
@@ -1,0 +1,5 @@
+import type { ReactNode } from 'react';
+
+export default function CourseIdLayout({ children }: { children: ReactNode }) {
+  return <>{children}</>;
+}
